requirements.txt
fastapi~=0.110.0
tortoise-orm~=0.20.0
slowapi~=0.1.9
starlette~=0.36.3
pydantic~=2.6.4
redis~=5.0.3
httpx~=0.27.0
loguru~=0.7.2
ossapi~=3.4.4
pillow~=10.2.0
numpy~=1.26.4
matplotlib~=3.8.3
pytz~=2024.1
